% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Candidate_circuits_construction_with_TAD}
\alias{Candidate_circuits_construction_with_TAD}
\title{Build candidate circuits}
\usage{
Candidate_circuits_construction_with_TAD(loaded_data, TAD_file_path)
}
\arguments{
\item{loaded_data}{The output from \code{Data_loading}}

\item{TAD_file_path}{The path to the TAD prior file}
}
\value{
A list named "Candidate_circuits" containing the following elements:
\itemize{
\item TFs: candidate TFs
\item TF_log2Count: log2 transformed TF expression
\item Peaks: candidate peaks
\item Peak_log2Count: log2 transformed peak accessibility
\item Genes: candidate genes
\item Gene_log2Count: log2 transformed gene expression
\item TF_peak_binding: TF-peak binding matrix
\item Peak_Gene_looping: peak-gene looping matrix
}
}
\description{
This function constructs candidate circuits based on the input data WITH TAD prior. See also \code{Candidate_circuits_construction_without_TAD}
}
