% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Data_loading}
\alias{Data_loading}
\title{Loading Data}
\usage{
Data_loading(
  Candidate_Gene_file_path,
  Candidate_Peak_file_path,
  scRNA_readcount_file_path,
  scRNA_Gene_file_path,
  scRNA_cellmeta_file_path,
  scATAC_readcount_file_path,
  scATAC_Peak_file_path,
  scATAC_cellmeta_file_path,
  Motif_mapping_file_path,
  Motif_name_file_path,
  Ref_seq_file_path
)
}
\arguments{
\item{scRNA_readcount_file_path, scRNA_gene_file_path, scRNA_cellmeta_file_path}{paths to filtered scRNA data of the cell type}

\item{scATAC_readcount_file_path, scATAC_peak_file_path, scATAC_cellmeta_file_path}{paths to filtered scATAC data of the cell type}

\item{Motif_mapping_file_path, Motif_name_file_path}{paths to TF motif prior on all ATAC peaks}

\item{Ref_seq_file_path}{path to the Refseq file for transcription starting site extraction}

\item{Candidate_gene_file_path, Candidate_peak_file_path}{paths to pre-selected candidate genes and peaks for the cell type}
}
\value{
A list named "loaded_data" containing the following elements:
\itemize{
\item Common_samples: a vector of common samples between scRNA and scATAC data
\item Candidate_Genes: a data frame of candidate genes
\item Candidate_Peaks: a data frame of candidate peaks
\item scRNA_Genes: a data frame of scRNA genes
\item scRNA_cells: a data frame of scRNA cells
\item scRNA_read_count_matrix: a sparse matrix of scRNA read counts
\item scATAC_Peaks: a data frame of scATAC peaks
\item scATAC_cells: a data frame of scATAC cells
\item scATAC_read_count_matrix: a sparse matrix of scATAC read counts
\item Motifs: a data frame of TF motifs
\item TF_peak_binding_matrix: a sparse matrix of TF binding on ATAC peaks
\item Refseq: a data frame of Refseq
}
}
\description{
Load data from files.
}
\seealso{
\code{\link[=Data_loading_from_workspace]{Data_loading_from_workspace()}}
}
