% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Data_loading_from_workspace}
\alias{Data_loading_from_workspace}
\title{Loading Data}
\usage{
Data_loading_from_workspace(
  Common_samples,
  Candidate_Genes,
  Candidate_Peaks,
  scRNA_Genes,
  scRNA_cells,
  scRNA_read_count_matrix,
  scATAC_Peaks,
  scATAC_cells,
  scATAC_read_count_matrix,
  Motifs,
  TF_Peak_binding_matrix,
  Refseq
)
}
\value{
A list named "loaded_data" containing the following elements:
\itemize{
\item Common_samples: a vector of common samples between scRNA and scATAC data
\item Candidate_Genes: a data frame of candidate genes
\item Candidate_Peaks: a data frame of candidate peaks
\item scRNA_Genes: a data frame of scRNA genes
\item scRNA_cells: a data frame of scRNA cells
\item scRNA_read_count_matrix: a sparse matrix of scRNA read counts
\item scATAC_Peaks: a data frame of scATAC peaks
\item scATAC_cells: a data frame of scATAC cells
\item scATAC_read_count_matrix: a sparse matrix of scATAC read counts
\item Motifs: a data frame of TF motifs
\item TF_peak_binding_matrix: a sparse matrix of TF binding on ATAC peaks
\item Refseq: a data frame of Refseq
}
}
\description{
Load data from workspace.
}
\seealso{
\code{\link[=Data_loading]{Data_loading()}} for loading data from files.
}
