% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MAGICAL_circuits_output}
\alias{MAGICAL_circuits_output}
\title{MAGICAL output}
\usage{
MAGICAL_circuits_output(
  Output_file_path,
  Candidate_circuits,
  Circuits_linkage_posterior,
  prob_threshold_TF_peak_binding = 0.8,
  prob_threshold_peak_gene_looping = 0.95
)
}
\arguments{
\item{Output_file_path}{The output file path}

\item{Candidate_circuits}{The output from "Candidate_circuits" (with or without TAD)}

\item{Circuits_linkage_posterior}{The output from \code{MAGICAL_estimation}}

\item{prob_threshold_TF_peak_binding}{The threshold of TF-peak binding probability. Default is 0.8}

\item{prob_threshold_peak_gene_looping}{The threshold of peak-gene looping probability. Default is 0.95}
}
\value{
A file with the format: Gene_symbol - Gene_chr - Gene_TSS - Peak_chr - Peak_start - Peak_end - Looping_prob - TFs(binding prob)
}
\description{
MAGICAL output
}
