% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MAGICAL_estimation}
\alias{MAGICAL_estimation}
\title{MCMC sampling}
\usage{
MAGICAL_estimation(
  loaded_data,
  Candidate_circuits,
  Initial_model,
  iteration_num
)
}
\arguments{
\item{loaded_data}{The output from \code{Data_loading}}

\item{Candidate_circuits}{The output from "Candidate_circuits" (with or without TAD)}

\item{Initial_model}{The output from \code{MAGICAL_initialization}}

\item{iteration_num}{The number of iterations}
}
\value{
A list named "Circuits_linkage_posterior" containing the following elements:
\itemize{
\item TF_Peak_Binding_prob: The posterior probability of TF binding to peaks
\item Peak_Gene_Looping_prob: The posterior probability of looping between peaks and genes
\item Noise_parameters: The posterior noise parameters
}
}
\description{
This function estimates the parameters.
}
