% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MAGICAL_initialization}
\alias{MAGICAL_initialization}
\title{MCMC model initialization}
\usage{
MAGICAL_initialization(loaded_data, Candidate_circuits)
}
\arguments{
\item{loaded_data}{The output from \code{Data_loading}}

\item{Candidate_circuits}{The output from "Candidate_circuits" (with or without TAD)}
}
\value{
A list of initial values for the MCMC model
}
\description{
This function initializes the MCMC model for MAGICAL.
}
\details{
For MCMC method, the initial values of parameters are not deterministic for the results.
Use pseudo bulk data to initialize the model as only at this dimension all data are matched.
}
