% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_circuits_with_idx}
\alias{plot_circuits_with_idx}
\title{Plot circuit(s)}
\usage{
plot_circuits_with_idx(data, idx, gene_track = T, TxDb, peak_track = T, peaks)
}
\arguments{
\item{data}{The data frame written by \code{\link[=MAGICAL_circuits_output]{MAGICAL_circuits_output()}}.}

\item{idx}{The index of the circuit to be plotted.}

\item{gene_track}{Whether to plot gene tracks around the region or not. Default is \code{TRUE}.}

\item{TxDb}{The TxDb object corresponding to the reference genome. Required if \code{gene_track} is \code{TRUE}.}

\item{peak_track}{Whether to plot peak tracks around the region or not. Default is \code{TRUE}.}

\item{peaks}{The data frame containing the peaks. It should contain 3 columns: the chromosome, start, and end of the peaks. Required if \code{peak_track} is \code{TRUE}.}
}
\value{
A karyoploteR plot.
}
\description{
This function plots one circuit by its index.
}
\examples{
# data = read.table("MAGICAL_selected_regulatory_circuits.txt", header = T, sep = "\t")
#
# library("TxDb.Hsapiens.UCSC.hg38.knownGene")
#
# peaks = read.table('Demo input files/scATAC peaks.txt', header = F, sep = "\t")
#
# plot_circuits_with_idx(data, 1, gene_track = T, TxDb.Hsapiens.UCSC.hg38.knownGene, peak_track = T, peaks)

}
\seealso{
\code{\link[=plot_circuits_with_gene]{plot_circuits_with_gene()}} and \code{\link[=plot_circuits_with_peak]{plot_circuits_with_peak()}}
}
