% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_circuits_with_idx}
\alias{plot_circuits_with_idx}
\title{Plot circuit(s)}
\usage{
plot_circuits_with_idx(data, idx, gene_track = F, TxDb)
}
\arguments{
\item{data}{The data frame written by \code{\link[=MAGICAL_circuits_output]{MAGICAL_circuits_output()}}.}

\item{idx}{The index of the circuit to be plotted.}

\item{gene_track}{Whether to plot gene tracks around the region or not. Default is \code{FALSE}.}

\item{TxDb}{The TxDb object corresponding to the reference genome. Required if \code{gene_track} is \code{TRUE}.}
}
\value{
A karyoploteR plot.
}
\description{
This function plots one circuit by its index.
}
\examples{
#library("TxDb.Hsapiens.UCSC.hg38.knownGene")
#data = read.table("MAGICAL_selected_regulatory_circuits.txt", header = T, sep = "\t")
#plot_circuits_with_idx(data, 1, T, TxDb.Hsapiens.UCSC.hg38.knownGene)

}
\seealso{
\code{\link[=plot_circuits_with_gene]{plot_circuits_with_gene()}} and \code{\link[=plot_circuits_with_peak]{plot_circuits_with_peak()}}
}
