% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_circuits_with_peak}
\alias{plot_circuits_with_peak}
\title{Plot circuit(s)}
\usage{
plot_circuits_with_peak(
  data,
  peak_chr,
  peak_start,
  peak_end,
  gene_track = F,
  TxDb
)
}
\arguments{
\item{data}{The data frame written by \code{\link[=MAGICAL_circuits_output]{MAGICAL_circuits_output()}}.}

\item{peak_chr}{(Character) the chromosome of the peak.}

\item{peak_start}{The start position of the peak.}

\item{peak_end}{The end position of the peak.}

\item{gene_track}{Whether to plot gene tracks around the region or not. Default is \code{FALSE}.}

\item{TxDb}{The TxDb object corresponding to the reference genome. Required if \code{gene_track} is \code{TRUE}.}
}
\value{
A karyoploteR plot.
}
\description{
This function plots all circuits containing one certain peak.
}
\examples{
#library("TxDb.Hsapiens.UCSC.hg38.knownGene")
#data = read.table("MAGICAL_selected_regulatory_circuits.txt", header = T, sep = "\t")
#plot_circuits_with_peak(data,"chr4", 184817550, 184818412, T, TxDb.Hsapiens.UCSC.hg38.knownGene)

}
\seealso{
\code{\link[=plot_circuits_with_idx]{plot_circuits_with_idx()}} and \code{\link[=plot_circuits_with_gene]{plot_circuits_with_gene()}}
}
